moved tooltip to better spot, auto time values
