+++ src/Ux/DateTimeRangeSelection.tsx
@@ -1,8 +1,9 @@
 /** @jsxImportSource @emotion/react */
+import { css } from "@emotion/react";
 import React from "react";
 import { Tooltip } from "@mui/material";
-import { TRACY_MAX_FILE_SIZE } from "../constants";
-import { formatNumber, parseDateString } from "../utility";
+import { DEFAULT_TIME_SELECTION_STEPSIZE, DEFAULT_TIME_SELECTION_STEPSIZE_CTRL, DEFAULT_TIME_SELECTION_STEPSIZE_SHIFT, TRACY_MAX_FILE_SIZE } from "../constants";
+import { formatNumber, msToTimeString, parseDateString } from "../utility";
 import { Ext2WebMessage, updateWebviewState, vscodeAPI } from "../communicationProtocol";
 import DateTimeRangeSlider from "./DateTimeRangeSlider";
 import FileTimeline from "./FileTimeline";
@@ -80,6 +81,25 @@ export default function DateTimeRangeSelection({ onDirtyMetadata }: Props) {
         
         <div css={{ display: 'flex', alignItems: 'flex-start' }}>
             <div css={{ marginRight: "10px" }}>
+                {/* Helper Tooltip */}
+                <Tooltip title={<div>
+                        <h2 css={{ fontSize: "16px", fontWeight: "bold", marginBottom: "2px" }}>Help</h2>
+                        <ul css={{ marginTop: "2px" }}>
+                            <li css={helpListItemStyle}>Default step size: <b>{msToTimeString(DEFAULT_TIME_SELECTION_STEPSIZE, false)}</b>.</li>
+                            <li css={helpListItemStyle}>Hold <b>Shift</b> for a step size of
+                                <b> {msToTimeString(DEFAULT_TIME_SELECTION_STEPSIZE / DEFAULT_TIME_SELECTION_STEPSIZE_SHIFT, false)}</b>.
+                            </li>
+                            <li css={helpListItemStyle}>Hold <b>Ctrl</b> for a step size of
+                                <b> {msToTimeString(DEFAULT_TIME_SELECTION_STEPSIZE / DEFAULT_TIME_SELECTION_STEPSIZE_CTRL, true)}</b>.
+                            </li>
+                            <li css={helpListItemStyle}>Hold <b>Shift</b> and <b>Ctrl</b> for a stepsize of
+                                <b> {msToTimeString(DEFAULT_TIME_SELECTION_STEPSIZE / (DEFAULT_TIME_SELECTION_STEPSIZE_SHIFT * DEFAULT_TIME_SELECTION_STEPSIZE_CTRL), true)}</b>.
+                            </li>
+                        </ul>
+                        <span css={{ fontSize: "14px" }}>Use the arrow keys to fine-tune the selected slider thumb.</span>
+                    </div>}>
+                    <i className="codicon codicon-question" />
+                </Tooltip>
                 <div>
                     <VSCodeButton disabled={dates.begin === topOfZoomStack[0] && dates.end === topOfZoomStack[1]}
                         onClick={() => { setZoomStack(zoomStack => zoomStack.concat([[ dates.begin, dates.end ]])) }}>
@@ -113,3 +133,6 @@ export default function DateTimeRangeSelection({ onDirtyMetadata }: Props) {
         </Tooltip>
     </div>);
 }
+
+const helpListItemStyle = css({ fontSize: "12px", padding: "2px", listStyleType: "circle"});
+
